+++ tests/Fixtures/UserResource.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace OptimistDigital\NovaDetachedFilters\Tests\Fixtures;
+namespace Outl1ne\NovaDetachedFilters\Tests\Fixtures;
 
 use Illuminate\Http\Request;
 use Laravel\Nova\Resource;
 use Laravel\Nova\Tests\Fixtures\User;
-use OptimistDigital\NovaDetachedFilters\HasDetachedFilters;
-use OptimistDigital\NovaDetachedFilters\NovaDetachedFilters;
+use Outl1ne\NovaDetachedFilters\HasDetachedFilters;
+use Outl1ne\NovaDetachedFilters\NovaDetachedFilters;
 
 class UserResource extends Resource
 {